+++ slurm/slurmdb.h
@@ -1286,11 +1286,24 @@ extern List slurmdb_txn_get(void *db_conn, slurmdb_txn_cond_t *txn_cond);
  * the position of the id in the select plugin array, as well as sets up the
  * control_addr and dim_size parts of the structure.
  *
- * IN: cluster_names - comman separated string of cluster names
+ * IN: cluster_names - comma separated string of cluster names
  * RET: List of slurmdb_cluster_rec_t *
  * note List needs to bbe freed with slurm_list_destroy() when called
  */
 extern List slurmdb_get_info_cluster(char *cluster_names);
+
+/*
+ * get the first cluster that will run a job
+ * IN: req - description of resource allocation request
+ * IN: cluster_names - comma separated string of cluster names
+ * OUT: cluster_rec - record of selected cluster or NULL if none found or
+ * 		      cluster_names is NULL
+ * RET: SLURM_SUCCESS on success SLURM_ERROR else
+ * note cluster_rec needs to be freed with slurmdb_destroy_cluster_rec() when
+ * called
+ */
+extern int slurmdb_get_first_avail_cluster(job_desc_msg_t *req,
+        char *cluster_names, slurmdb_cluster_rec_t **cluster_rec);
 
 /************** helper functions **************/
 extern void slurmdb_destroy_user_defs(void *object);
