+++ src/sbatch/Makefile.am
@@ -7,7 +7,7 @@ AM_CPPFLAGS = -I$(top_srcdir)
 
 bin_PROGRAMS = sbatch
 
-sbatch_SOURCES = sbatch.c mult_cluster.c mult_cluster.h opt.c opt.h
+sbatch_SOURCES = sbatch.c opt.c opt.h
 
 convenience_libs = $(top_builddir)/src/api/libslurm.o $(DL_LIBS)
